% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_cases_for_square.R
\name{compute_square}
\alias{compute_square}
\title{@title compute_square}
\usage{
compute_square(x)
}
\arguments{
\item{x}{this is a real number}
}
\value{
the square of a number
}
\description{
@title compute_square
}
\details{
: this will return the square of any number called x
}
\examples{
x <- 2
          compute_square(x)


}
