% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_proof.r
\name{error_proof}
\alias{error_proof}
\title{error_proof}
\usage{
error_proof(x, return_na = FALSE)
}
\arguments{
\item{x}{The function/code chunk you which to safely run}

\item{return_na}{Logical indicating whether ot not to return NA upon error.}
}
\value{
The output of the code chunk if successful, or NA if an error occurs.
}
\description{
This code allows you to run code safely within a loop or vectorised
without errors breaking the operation. Should an error occur, the error message will be
printed to the console and NA returned.
}
