% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_file_paths.R
\name{list_file_paths}
\alias{list_file_paths}
\title{list_file_paths}
\usage{
list_file_paths(
  path,
  pattern = NULL,
  all.files = FALSE,
  full.names = FALSE,
  recursive = FALSE,
  ignore.case = FALSE,
  include.dirs = FALSE,
  no.. = FALSE,
  return_absolute = FALSE
)
}
\arguments{
\item{path}{a character vector of full path names; the default corresponds to the working directory, getwd(). Tilde expansion (see path.expand) is performed. Missing values will be ignored.}

\item{pattern}{an optional regular expression. Only file names which match the regular expression will be returned.}

\item{all.files}{a logical value. If FALSE, only the names of visible files are returned. If TRUE, all file names will be returned.}

\item{full.names}{a logical value. If TRUE, the directory path is prepended to the file names to give a relative file path. If FALSE, the file names (rather than paths) are returned.}

\item{recursive}{logical. Should the listing recurse into directories?}

\item{ignore.case}{logical. Should pattern-matching be case-insensitive?}

\item{include.dirs}{logical. Should subdirectory names be included in recursive listings? (They always are in non-recursive ones).}

\item{no..}{logical. Should both "." and ".." be excluded also from non-recursive listings?}

\item{return_absolute}{logical. Should absolute file paths be returned. If false, only relative file paths are returned.}
}
\value{

}
\description{
a wrapper for the base function list.files(). Instead of returning just the list of the file names, it returns the full path to those files as a named list.
}
