% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/not_in.r
\name{not_in}
\alias{not_in}
\title{not_in}
\usage{
not_in(x, y)
}
\arguments{
\item{x}{A vector.}

\item{y}{A comparison vector.}
}
\value{
A logical vector of equal length to x indicating which elements of x are not located in y.
}
\description{
Checks for where x is not in y. The opposite of the base %in% function.
}
