% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normal_curve_shade.R
\name{the_normal_curve}
\alias{the_normal_curve}
\title{normal_curve_shade}
\usage{
the_normal_curve(
  x,
  mean,
  sd,
  col,
  VerticalSplit,
  shadeFrom,
  shadeTo,
  plot_type = "base",
  ...
)
}
\arguments{
\item{x}{a sequence of the range of the x-axis}

\item{mean}{this is the mean defined for the normal curve}

\item{sd}{this is the standard deviation defined for the normal curve}

\item{col}{this is the color for the shaded area}

\item{VerticalSplit}{the point where the x-axis should split vertically}

\item{shadeFrom}{The point on the x-axis where the shade should start from}

\item{shadeTo}{The point on the x-axis where the shade should end at}

\item{plot_type}{specifies the type of plot if it is base of ggplot}

\item{...}{the is for any other parameters not predefined but can be defined in a plot function}
}
\value{
No returns, just creates a plot
}
\description{
normal_curve_shade
}
\details{
: This is the function that performs a normal curve and also shade the a specified area
}
\examples{
x <- seq(1.5, 10.5, 0.05 )
          the_normal_curve(x=x, mean = 6, sd = 2, col = "blue", VerticalSplit = 7, shadeFrom = 7, shadeTo = 10.5, plot_type = "ggplot" ...)
}
