% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparkplot.R
\name{the_spark_plot}
\alias{the_spark_plot}
\title{the_spark_plot Just Another Spark-line plot function}
\usage{
the_spark_plot(dat, trans, col, label_x, label_y, the_title, ...)
}
\arguments{
\item{dat}{the data that has to be a dataframe}

\item{trans}{level of transparency of the line}

\item{col}{colour of the plot}

\item{label_x}{label name for the x-axis}

\item{label_y}{label name for the y-axis}

\item{the_title}{title of the plot}
}
\value{
No returns, just creates a plot
}
\description{
the_spark_plot Just Another Spark-line plot function
}
\details{
: takes a dataframe and turns it into a sparkline plot where only the first line is not transparent to a specified opacity level
}
\examples{
dat <- data.frame(matrix(runif(50*3, 0, 1 ), 50))
          the_spark_plot(dat=dat, trans = 0.3, col = "red", label_x = "Data point", label_y = "Value", the_title = "Sparkline Plot")

}
