% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weighted_se.r
\name{weighted_se}
\alias{weighted_se}
\title{weighted_se}
\usage{
weighted_se(x, w, na.rm = TRUE)
}
\arguments{
\item{x}{A numeric vector}

\item{w}{A numeric vector of weights}

\item{na.rm}{Logical indicating wheter NA's should be removed.}
}
\value{
The weighted standard error of x.
}
\description{
Calculate weighted standard error of x using weights w.
}
