% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weighted_var.r
\name{weighted_var}
\alias{weighted_var}
\title{weighted_var}
\usage{
weighted_var(x, w = NULL, na.rm = FALSE)
}
\arguments{
\item{x}{A numeric vector.}

\item{w}{A numeric vector to weight x by.}

\item{na.rm}{Logical indicating whether to ignore NAs.}
}
\value{
The weighted variance of x.
}
\description{
Calculate weighted variance of x using weights w.
}
