% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_excel_allsheets.r
\name{read_excel_allsheets}
\alias{read_excel_allsheets}
\title{read_excel_allsheets}
\usage{
read_excel_allsheets(file_path, tibble = FALSE)
}
\arguments{
\item{file_path}{path to the excel file.}

\item{tibble}{logical indicating whether to return a tibble or a dataframe.}
}
\value{
A list of tibbles or data frames
}
\description{
function for reading in all sheets in an excel workbook
}
